requirements.txt
Django==1.7
PyJWT==1.1.0
South==1.0
argparse==1.2.1
boto==2.36.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-ckeditor==4.4.7
django-compressor==1.4
django-debug-toolbar==1.2.2
django-filter==0.9.1
django-google-oauth==1.0.4
django-grappelli==2.6.3
django-gravatar2==1.3.0
django-password-reset==0.7
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.0.0
djrill==1.3.0
drf-nested-routers==0.9.0
google-api-python-client==1.3.1
gunicorn==19.1.1
httplib2==0.9
mock==1.0.1
newrelic==2.40.0.34
oauth2client==1.4.5
oauthlib==0.7.2
psycopg2==2.5.4
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-googleanalytics==1.0.2
python-openid==2.2.5
python-social-auth==0.2.7
requests==2.5.0
requests-oauthlib==0.4.2
rsa==3.1.4
selenium==2.44.0
sendgrid==1.2.3
simplejson==3.6.5
six==1.8.0
smtpapi==0.1.3
sqlparse==0.1.13
static3==0.5.1
unittest2==0.8.0
uritemplate==0.6
wsgiref==0.1.2
